Add the testing files.
